requirements.txt
utils-nuuuwan
camelot-py[cv]
matplotlib
gig-nuuuwan
pymupdf